ReactNodeJS/Internship/python_backend/requirements.txt
elasticsearch
datetime
nltk
spacy
jsonify
Flask
google-cloud-storage
google-cloud
google-cloud-language
google-api-python-client
